big_brother_cnn/requirements.txt
torch>=1.13.0
torchvision>=0.14.0
numpy>=1.21.0
pandas>=1.3.0
opencv-python>=4.5.0
scikit-learn>=1.0.0
pyyaml>=5.4.0
python-dateutil>=2.8.0
pytz>=2021.1
pillow>=8.3.0
matplotlib>=3.5.0
tqdm>=4.64.0 